Move Gradle dependencies out of app-engine
